Move blockIndex replacement back to onSubmit
Due to a change to core JS that moved blockIndex replacment from immediately before submit to block creation, we need to replace blockIndex here. Otherwise, dynamically created WMS inputs are ignored. Unfortunately the only way to get the blockIndex at this stage is to extract it from the layout input's name (ideally the index would be set to a data attribute, but that doesn't exist at the time of this fix).