Detach pending mapping entities, too
The event listener we use to fix the situation where a Mapping entity
refers to a detached/non-existent Item (or Media) doesn't work in the
case where the mapping entity is _new_.

We need to look in the pending insertions from the Doctrine unit of work
in that case.

This problem manifests as errors during imports when an item isn't
actually persisted due to an error (the concrete example here being an
Omeka2Importer import where some items were not created due to a mime
type issue for their files; stray Mapping entities converted those
skipped items into a fatal error.