Fix round-trip hydration for markers
Markers are output to the API as references, but input assumes the basic
format of the admin UI form, in particular that the location data will
come in with every request. This means that non-patch requests are
interpreted as attempts to keep a marker but clear it of all data,
leading to a validation error.

The problem occurs when using the API output as input, a typical
pattern for external clients, and also occurs for other modules that use
this pattern like CSV Import.

This commit solves the round-trip problem for markers by treating a
"bare" marker array (only the ID keys specified) as simply a request to
retain the marker as-is.

(#97, #98)