Fix round-trip hydration for mappings
Mappings have a slightly different problem than markers: default
round-trip hydration for them just removes the mapping bounds rather
than causing an error. The error caused by marker hydration would have
masked this for all typical cases, though.

The basic nature of the fix here is the same as for markers: skip doing
any updating if the passed "mapping data" is just the default reference.
The one-to-one nature of mappings, though, leads to a more significant
change here. Rather than reading the passed-in mapping ID as before,
this change shifts to always operating on the mapping entity associated
with the item being updated, which simplifies some logic and prevents
some possible errors (previously it was possible to cause Mapping to
attempt to create a new mapping entity by omitting the ID; if there was
one already this leads to a foreign key constraint error at the SQL
level).

(#97 #98)