Fix hydration for mappings
Mappings handling relied on the API client to pass the correct ID for
the mapping entity with all requests. Omitting the ID when a mapping
entity existed for the item caused the module to try to make a new one,
hitting an error on the SQL-level unique constraint on the table. Giving
the wrong ID could update/delete the mapping data for a different item.

This commit fixes this behavior by ignoring the passed-in mapping ID,
instead determining the present mapping data by looking up the mapping
(if any) for the current item.

(#97 #98)