Always skip empty cells
The lat-alone and lng-alone mappings would always create a marker for
every row, even where the values for the cells were blank. Combined
lat/lng was immune to this just because of how its method of splitting
the value worked.

This makes the CSV mapping never set any mapping data for a blank cell.